visualizations_dash/requirements.txt
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.18.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.8.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.2.0
patsy==0.5.1
plotly==4.14.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
retrying==1.3.3
scipy==1.6.0
Shapely==1.7.1
six==1.15.0
statsmodels==0.12.1
threadpoolctl==2.1.0
urllib3==1.26.2
Werkzeug==1.0.1
